sigie: fix perror messages
The function perror adds a description of the latest happened error to the
user-specified message. The description follows the symbol ':', thus user's
message should not contain any '.' or '\n' at the end.

The commit removes all unappropriate symbols from ends of user's messages. It
also adds the key "[error]" at the beginning to indicate that the message is
about some inappropriate behavior in the program.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>