sigie: fix lookup through the variable list
The end of the variable list was determined incorrectly using a pointer to the
data and a counter of bytes used in the buffer. This commit fixes the problem by
correct pointer arithmetics to calculate the top boundary of variables memory
during variable value lookup and printing of all variables.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>